# Conflicts:
#	src/integration-test/kotlin/de/bixilon/minosoft/MinosoftSIT.kt
#	src/integration-test/kotlin/de/bixilon/minosoft/data/physics/PhysicsTestUtil.kt
#	src/integration-test/kotlin/de/bixilon/minosoft/protocol/network/connection/play/ConnectionTestUtil.kt
#	src/main/java/de/bixilon/minosoft/data/registries/blocks/properties/BlockProperties.kt
#	src/main/java/de/bixilon/minosoft/data/registries/blocks/types/pixlyzer/PixLyzerBlock.kt
#	src/main/java/de/bixilon/minosoft/gui/rendering/models/block/state/baked/BakedFace.kt
#	src/main/java/de/bixilon/minosoft/gui/rendering/system/base/texture/dynamic/DynamicTextureListener.kt
#	src/main/java/de/bixilon/minosoft/protocol/protocol/ProtocolVersions.kt
#	src/main/java/de/bixilon/minosoft/util/collections/floats/FragmentedArrayFloatList.kt
#	src/main/resources/assets/minosoft/mapping/assets_properties.json
#	src/main/resources/assets/minosoft/mapping/versions.json

+++ src/main/java/de/bixilon/minosoft/data/entities/entities/LightningBolt.kt
@@ -13,7 +13,7 @@
 package de.bixilon.minosoft.data.entities.entities
 
 import de.bixilon.kotlinglm.vec3.Vec3d
-import de.bixilon.kutil.random.RandomUtil.nextLong
+import de.bixilon.kutil.random.RandomUtil.nextInt
 import de.bixilon.minosoft.data.entities.EntityRotation
 import de.bixilon.minosoft.data.entities.data.EntityData
 import de.bixilon.minosoft.data.registries.entities.EntityFactory
@@ -21,10 +21,9 @@ import de.bixilon.minosoft.data.registries.entities.EntityType
 import de.bixilon.minosoft.data.registries.identified.Namespaces.minecraft
 import de.bixilon.minosoft.data.registries.identified.ResourceLocation
 import de.bixilon.minosoft.protocol.network.connection.play.PlayConnection
-import de.bixilon.minosoft.util.KUtil
 
 class LightningBolt(connection: PlayConnection, entityType: EntityType, data: EntityData, position: Vec3d) : Entity(connection, entityType, data, position, EntityRotation(0.0f, 0.0f)) {
-    val duration = random.nextLong(100, 250)
+    val duration = random.nextInt(100, 250)
 
     companion object : EntityFactory<LightningBolt> {
         override val identifier: ResourceLocation = minecraft("lightning_bolt")
